Вынес логику добавления todo в компонент TodoList
