+++ src/components/TodoList.tsx
@@ -1,25 +1,113 @@
 import { TodoItem, Todo } from './TodoItem';
-import React from 'react';
-interface Todos {
-  todos: Todo[];
-  onDelete: (id: number) => void;
-  onChanging: (id: number, newTodoName: string) => void;
-  onChangeCompleted: (id: number, completed: boolean) => void;
-}
-export const TodoList = ({ todos, onDelete, onChanging, onChangeCompleted }: Todos) => {
+import React, { useMemo, useState, useEffect } from 'react';
+import styled from 'styled-components';
+import { Button, Input } from 'antd';
+import { Card } from './Card';
+
+export const TodoList = () => {
+  function onDelete(id: number) {
+    setTodos((prevTodos) =>
+      prevTodos.filter((prevTodo) => {
+        return prevTodo.id !== id;
+      }),
+    );
+  }
+
+  function onChanging(id: number, newTodoName: string) {
+    setTodos((prevTodos) =>
+      prevTodos.map((prevTodo) => {
+        if (prevTodo.id === id)
+          return { ...prevTodo, changing: !prevTodo.changing, name: prevTodo.changing ? newTodoName : prevTodo.name };
+        return prevTodo;
+      }),
+    );
+  }
+
+  async function onChangeCompleted(id: number, completed: boolean) {
+    await setTodos((prevTodos) =>
+      prevTodos.map((prevTodo) => {
+        if (prevTodo.id === id) return { ...prevTodo, completed };
+        return prevTodo;
+      }),
+    );
+  }
+
+  function filterTodoList(searchWord: string, todoList: Todo[]) {
+    const lowerSearchWord = searchWord.toLowerCase();
+    return todoList.filter((todo) => {
+      return todo.name.toLowerCase().includes(lowerSearchWord);
+    });
+  }
+
+  async function addTodo(newTodo: Todo) {
+    await setTodos((prevTodos) => {
+      return [...prevTodos, newTodo];
+    });
+    setTempValue('');
+  }
+
+  function getTodosFromLocalStorage(): Todo[] | undefined {
+    return JSON.parse(localStorage.getItem('todos') as string);
+  }
+
+  const [tempValue, setTempValue] = useState('');
+  const [todos, setTodos] = useState<Todo[]>(
+    getTodosFromLocalStorage() || [
+      { id: 1, name: 'todo1', completed: false, changing: true },
+      { id: 2, name: 'todo2', completed: true, changing: false },
+      { id: 3, name: 'todo3', completed: false, changing: false },
+    ],
+  );
+
+  const filteredTodoList = useMemo<Todo[]>(() => {
+    return filterTodoList(tempValue, todos);
+  }, [tempValue, todos]);
+
+    useEffect(() => {
+        localStorage.setItem('todos', JSON.stringify(todos));
+    });
+
   return (
-    <div>
-      {todos.map((todoItem) => {
-        return (
-          <TodoItem
-            onDelete={() => onDelete(todoItem.id)}
-            onChanging={onChanging}
-            onChangeCompleted={onChangeCompleted}
-            {...todoItem}
-              key={todoItem.id}
+    <>
+      <Card>
+        <CreatingTodoPart>
+          <Input
+            style={{ width: '250px' }}
+            value={tempValue}
+            onChange={(e) => {
+              setTempValue(e.target.value);
+            }}
           />
-        );
-      })}
-    </div>
+          <Button
+            type="primary"
+            onClick={() => {
+              if (tempValue.length)
+                addTodo({ name: tempValue, completed: false, id: Number(Date.now()), changing: false });
+            }}
+          >
+            Add
+          </Button>
+        </CreatingTodoPart>
+      </Card>
+      <div>
+        {filteredTodoList.map((todoItem) => {
+          return (
+            <TodoItem
+              onDelete={() => onDelete(todoItem.id)}
+              onChanging={onChanging}
+              onChangeCompleted={onChangeCompleted}
+              {...todoItem}
+              key={todoItem.id}
+            />
+          );
+        })}
+      </div>
+    </>
   );
 };
+
+const CreatingTodoPart = styled.div`
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
+`;
